Restructure code for ToricVarieties
