+++ src/ToricVarieties/CyclicQuotientSingularities/CyclicQuotientSingularities.jl
